refactor(diagrams,extension): single process-blueprint SVG body (review C, process-blueprint slice)
The VS Code preview and the host-neutral package renderer had drifted into two
different blueprints: the package output used opacity-based row fills,
single-line pills (`name · id`), non-interactive chips and no clip path, while
the editor rendered the richer layered body (clipped row fills + grid, on-top
pills/chips, interactive `<g data-chip-law>` chips, multi-line pills with a
separate id line, outer `bp-border`).

Unify on the VS Code body as canonical:

- Extract `renderProcessBlueprintBody(layout, ox, oy)` into
  `render-process-blueprint.ts` — the single emitter both hosts call.
- `renderProcessBlueprintSvg` (JCEF/IntelliJ) becomes a thin wrapper: lay out,
  simple `text-header` title, call the shared body, embed the shared theme CSS
  plus the blueprint-specific `bp-row-bg` / `bp-border` rules so the output
  stays self-contained.
- `process-blueprint-preview.ts` drops its inline `truncate` / `textCellSvg` /
  `complianceChipSvg` and the `layoutToSvg` body; `layoutToSvg` now wraps the
  shared body with the VS Code title block.

VS Code output is byte-identical (verified with a temporary parity test over a
synthetic layout exercising headers, legend, multi-line goal/result cells,
pills with and without an id line, every compliance-chip decoration, and grid
lines, at two canvas offsets). The IntelliJ/JCEF output changes intentionally
to adopt the richer editor rendering — visual review expected.
